Removed inline declarations of "magic"
The inline declarations of the "magic" union have been replaced by a
single declaration in types.h called signedness_switch_trick. The
gen_assm.c file has also been changed to not have a declaration after a
label, which is apparently only allowed in C23. Also in gen_assm.c, the
switch statement has been altered to have a case for NONE, which does
nothing. This suppresses the clang warning, and makes the lack of
behavior explicit.